+++ components/weather/index.ts
@@ -1,2 +1,4 @@
 export { CurrentWeather } from "./CurrentWeather";
 export { WeatherIcon, getWeatherIcon } from "./WeatherIcon";
+export { WeatherDayCard } from "./WeatherDayCard";
+export { WeatherForecast } from "./WeatherForecast";